move classes in lib/common to `azure-toolkit-common-lib`
