+++ azure-toolkit-libs/azure-toolkit-springcloud-lib/src/main/java/com/microsoft/azure/toolkit/lib/springcloud/model/SpringCloudClusterEntity.java
@@ -24,7 +24,7 @@ package com.microsoft.azure.toolkit.lib.springcloud.model;
 
 import com.microsoft.azure.management.appplatform.v2020_07_01.implementation.ServiceResourceInner;
 import com.microsoft.azure.management.appplatform.v2020_07_01.implementation.SkuInner;
-import com.microsoft.azure.toolkit.lib.common.IAzureEntity;
+import com.microsoft.azure.toolkit.lib.common.entity.IAzureEntity;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
