+++ azure-toolkit-libs/azure-toolkit-springcloud-lib/src/main/java/com/microsoft/azure/toolkit/lib/springcloud/model/SpringCloudDeploymentEntity.java
@@ -25,7 +25,7 @@ package com.microsoft.azure.toolkit.lib.springcloud.model;
 import com.microsoft.azure.management.appplatform.v2020_07_01.DeploymentInstance;
 import com.microsoft.azure.management.appplatform.v2020_07_01.DeploymentResourceStatus;
 import com.microsoft.azure.management.appplatform.v2020_07_01.implementation.DeploymentResourceInner;
-import com.microsoft.azure.toolkit.lib.common.IAzureEntity;
+import com.microsoft.azure.toolkit.lib.common.entity.IAzureEntity;
 import lombok.Getter;
 
 import javax.annotation.Nonnull;
